permit the use of arrays as tuples
It's not currently safe to use arrays as tuples due to a bug in the
circular-reference detection in _determineActualTypes. Given a value
such as ["abc", Right(123)] in an environment containing both Array
and Pair (the type comprising all two-element arrays), the function
is applied recursively to the nested values once for Array and once
for Pair. This is fine in the case of "abc"; for Right(123), though,
this is problematic. The first time, Right(123) is appended to $seen.
The second time, the identical object is found in $seen, erroneously
indicating a circular reference.

This commit removes the mutation.