underline invalid record value rather than entire record
When we determine that a value is not a member of the specified record
type, we do not currently make the reason clear: we underline the type's
entire string representation even if just one field is invalid.

This problem compounds with nesting, since the outermost record type is
underlined regardless of the depth at which validation fails.

This commit updates the "underline" function to draw attention to the
invalid field.

Invalid position.x, before:

    { color :: String, position :: { x :: Number, y :: Number } }
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Invalid position.x, after:

    { color :: String, position :: { x :: Number, y :: Number } }
                                          ^^^^^^