ensure that _determineActualTypes returns an array of unique types
Any environment consisting of at least one nullary type and at least
one non-nullary type describes an infinite number of concrete types.
[$.Array($.Unknown), $.Number], for example, describes:

  - Number
  - Array Number
  - Array (Array Number)
  - Array (Array (Array Number))
  - ...

It's thus impossible to expand the "unknowns" in the environment to
produce a finite set of concrete types (which could then be filtered).
Instead, the environment is at once filtered and expanded (whenever an
"unknown" is encountered). This may produce duplicate concrete types.

This commit updates _determineActualTypes to discard duplicate types.