prevent unnecessary recursive applications of _determineActualTypes
This commit fixes an error in 2f94dbc which resulted in unnecessary
recursive applications of _determineActualTypes when determining the
types of a collection whose elements are members of a unary type.

For a given collection c whose elements are members of some unary type,
the complexity of the faulty algorithm is O(t^n), where t is the number
of types of which all the inner values of c's elements are members, and
n is the number of elements of c.

The actual types of the first element of c should be determined, once,
then refined by interrogating the remaining elements of c. The faulty
algorithm, though, would determine the actual types of elements of c
t^1 + t^2 + t^3 + ... + t^n times. For example:

    c = [[1], [2], [3], [4], [5]]
    t = 3 (Number, ValidNumber, FiniteNumber)
    n = 5

    3^1 + 3^2 + 3^3 + 3^4 + 3^5

The superfluous recursion produced duplicate results, prompting the
deduplication added in 437f0a4.

This commit prevents superfluous recursion, obviating the need for
deduplication. The complexity is now O(tn), a dramatic improvement!