dramatically improve performance of UnaryType and BinaryType
The predicate and extractor functions provided to the UnaryType and
BinaryType type constructors play an important role in type checking.
Each predicate function may be applied hundreds of times when a small
program is run. This is not a problem, in and of itself.

If these oft-applied functions are themselves type-checked, though,
run-time type checking is significantly more expensive as a result.

This commit disables type checking of predicate and extractor functions.