make sure that sanctuary-def will run without "util" module
Fix so sanctuary-def can be used in CommonJS enviroments that does
not have the `util` module, such as scrimba.com.

Here is an example with sanctuary-def 0.22.0:
https://scrimba.com/scrim/cQGBN6cM

The def test "returns a function with "inspect" and "@@show" methods"
fails with the default `util` object but I guess you already know
that.